Do not add -beta.x.x versions, automagically generate index.html

+++ scripts/build-js.sh
@@ -1,15 +1,19 @@
-source `dirname $0`/_vars.sh
-
-# php "$REPO_PATH/doctum.phar" update doctum-config.php -v
+source $(dirname $0)/_vars.sh
 
 branches=()
 eval "$(git for-each-ref --shell --format='branches+=(%(refname))' refs/heads/)"
 
-generate () {
-    echo "$style - $1$reset"
+echo "${style}Generating $reset"
 
+generate () {
     ref=$1
     path="$REPO_PATH/docs/js/$ref"
+
+    if [ "$2" == "true" ] && [ -d "$path" ]; then
+        return
+    fi
+
+    echo "$style - $ref$reset"
 
     rm -rf $path
     mkdir $path
@@ -18,7 +22,7 @@ generate () {
     cp "$REPO_PATH/src/readme-js.md" "$path/README.md"
 
     cd $path
-    (cd $FLARUM_PATH && git checkout -q $ref)
+    (cd $FLARUM_PATH && git checkout -q -- . && git checkout -q $ref)
     npx esdoc -c esdoc.json 2>&1 >/dev/null
 
     cd $REPO_PATH
@@ -27,10 +31,15 @@ generate () {
 generate master
 generate mithril-2-update
 
-tags=$(cd $FLARUM_PATH && git tag | sort -V)
+tags=$(cd $FLARUM_PATH && git tag)
 
 for tag in $tags; do
-    generate $tag
+    if [[ "$tag" =~ -beta.[0-9]{1,}.[0-9]{1,}$ ]]; then
+        continue
+    fi
+
+    generate $tag true
 done
 
-bash $SCRIPTS_PATH/set-redirects.sh php
+bash $SCRIPTS_PATH/set-redirects.sh js
+bash $SCRIPTS_PATH/set-index-file.sh js
